implemented some additional styles to events part of a project, help for Anya is completed
