+++ src/components/events/eventPost/EventPost.jsx
@@ -5,6 +5,7 @@ import {withRouter} from 'react-router';
 import EventForm from "../../forms/EventForm/EventForm";
 import {authService} from "../../../services/auth.service";
 import {Role} from "../../../helpers/FakeAPI/Role";
+import './EventPost.scss';
 
 class EventPost extends Component {
 
@@ -31,21 +32,21 @@ class EventPost extends Component {
     render() {
         const { event, isModalShowed } = this.state;
         return (
-            <div className="event-page">
+            <div className="event-post">
                 {event ?
                     <>
-                        <h3 className="events-item__title">{event.name}</h3>
-                        <p className="events-item__text">Date: {event.date}</p>
+                        <h3 className="event-post__title">{event.name}</h3>
+                        <p className="event-post__text">Date: {event.date}</p>
                         {event.isFullDayEvent ?
                             <strong>FULL DAY EVENT</strong>
                             :
                             <>
-                                <p className="events-item__text">Start Time: {event.startTime}</p>
-                                <p className="events-item__text">End Time: {event.endTime}</p>
+                                <p className="event-post__text">Start Time: {event.startTime}</p>
+                                <p className="event-post__text">End Time: {event.endTime}</p>
                             </>
                         }
                         {authService.currentUser.role === Role.Admin &&
-                        <div className="event-page__btns">
+                        <div className="event-post__btns">
                             <Button
                                 buttonColorScheme="primary"
                                 buttonSize="small"
