+++ src/components/events/eventPost/EventPost.scss
@@ -1,0 +1,22 @@
+.event-post {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  padding: 40px 0;
+  border: 1px solid  #ccc;
+  &__title {
+    color: #d64db8;
+    font-size: 28px;
+  }
+  &__text {
+    font-weight: bold;
+    font-size: 23px;
+    margin: 10px;
+  }
+  &__btns {
+    button {
+      margin-left: 15px;
+    }
+  }
+}
