+++ src/components/events/eventsContainer/Events.scss
@@ -1,0 +1,8 @@
+.events {
+  &__add-btn {
+    width: 100%;
+    display: flex;
+    justify-content: flex-end;
+    padding: 20px;
+  }
+}
