+++ src/components/events/eventsItem/EventsItem.scss
@@ -2,11 +2,15 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    
+    border: 1px solid #ccc;
+    padding: 20px;
+
     &__title{
-        color: darkblue;
+        color: #d64db8;
+        font-size: 28px;
     }
     &__text {
         font-size: 22px;
+        margin: 10px;
     }
 }
